Calculation with criteria

+++ server/DecisionMakingServer/APIModels/ResultDTO.cs
@@ -7,5 +7,5 @@ public class ResultDTO
     [JsonProperty("rankingId")] public int RankingId { get; set; }
     [JsonProperty("criterionId")] public int CriterionId { get; set; }
     [JsonProperty("alternativeId")] public int AlternativeId { get; set; }
-    [JsonProperty("score")] public float Score;
+    [JsonProperty("score")] public double Score;
 }
